1.新增LICENSE文件 2.修复数据类型

+++ WPLogOutputerDemo/WPLogOutputer/WPLogOutputer.h
@@ -76,7 +76,7 @@
 /**
  log文件大小
  */
-+ (long)logFileSize;
++ (long long)logFileSize;
 @end
